[add] username queries & [feature] order CRUD

+++ src/main/java/com/beerproducts/backend_products_ms/controllers/ProductController.java
@@ -35,17 +35,32 @@ import lombok.AllArgsConstructor;
 public class ProductController {
     private final ProductService productService;
 
-    @GetMapping("/products")
-    public List<Product> fetchAllProducts() {
-        return productService.getAllProducts();
+    @GetMapping("/products/{username}/product/{id}")
+    public ResponseEntity<Product> getProductsById(@PathVariable String username, @PathVariable String id) {
+        List<Product> products = productService.findProductsByUsername(username);
+
+        if (products.size() > 0) {
+            Product product = productService.findProductById(id)
+                    .orElseThrow(() -> new ResourceNotFound("Not found products with the id " + id));
+
+            if (products.contains(product)) {
+                return new ResponseEntity<>(product, HttpStatus.OK);
+            }
+        }
+
+        throw new ResourceNotFound("Not found products with the username " + username);
+
     }
 
-    @GetMapping("/products/{id}")
-    public ResponseEntity<Product> fetchProduct(@PathVariable("id") String id) {
-        Product product = productService.getProductById(id)
-                .orElseThrow(() -> new ResourceNotFound("Not found product with id: " + id));
+    @GetMapping("/products/{username}")
+    public ResponseEntity<List<Product>> getProductsByUsername(@PathVariable String username) {
+        List<Product> products = productService.findProductsByUsername(username);
 
-        return new ResponseEntity<Product>(product, HttpStatus.OK);
+        if (products.size() > 0) {
+            return new ResponseEntity<>(products, HttpStatus.OK);
+        }
+
+        throw new ResourceNotFound("Not found products with the username " + username);
 
     }
 
@@ -60,31 +75,53 @@ public class ProductController {
         }
     }
 
-    @PutMapping("/products/{id}")
-    public ResponseEntity<Product> updateProduct(@Valid @RequestBody Product product, @PathVariable("id") String id) {
-        Product _product = productService.getProductById(id)
-                .orElseThrow(() -> new ResourceNotFound("Not found product with the id " + id));
+    @PutMapping("/products/{username}/update")
+    public ResponseEntity<Product> updateProduct(@Valid @RequestBody Product product, @PathVariable String username) {
+        List<Product> products = productService.findProductsByUsername(username);
 
-        _product.setName(product.getName());
-        _product.setDescription(product.getDescription());
-        _product.setImage(product.getImage());
-        _product.setPrice(product.getPrice());
-        _product.setStyle(product.getStyle());
-        _product.setCategory(product.getCategory());
-        _product.setAvg_grade(product.getAvg_grade());
-        _product.setIbu_grade(product.getIbu_grade());
+        if (products.size() > 0) {
 
-        return new ResponseEntity<>(productService.saveOrUpdateProduct(_product), HttpStatus.OK);
+            Product _product = productService.findProductById(product.getId())
+                    .orElse(null);
+
+            if (_product == null) {
+                throw new ResourceNotFound("Not found product with the id " + product.getId());
+            }
+
+            if (products.contains(_product)) {
+                _product.setName(product.getName());
+                _product.setDescription(product.getDescription());
+                _product.setImage(product.getImage());
+                _product.setPrice(product.getPrice());
+                _product.setStyle(product.getStyle());
+                _product.setCategory(product.getCategory());
+                _product.setAvg_grade(product.getAvg_grade());
+                _product.setIbu_grade(product.getIbu_grade());
+
+                return new ResponseEntity<>(productService.saveOrUpdateProduct(_product), HttpStatus.OK);
+            }
+        }
+        throw new ResourceNotFound("Not found products with the username " + username);
+
     }
 
-    @DeleteMapping("/products/{id}")
-    public ResponseEntity<Product> deleteProduct(@PathVariable("id") String id) {
-        Product product = productService.getProductById(id)
-                .orElseThrow(() -> new ResourceNotFound("Not found product with the id " + id));
+    @DeleteMapping("/products/{username}/delete/{id}")
+    public ResponseEntity<Product> deleteProduct(@PathVariable String id, @PathVariable String username) {
+        List<Product> products = productService.findProductsByUsername(username);
 
-        productService.deleteProductById(product.getId());
+        if (products.size() > 0) {
 
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            Product product = productService.findProductById(id)
+                    .orElseThrow(() -> new ResourceNotFound("Not found product with the id " + id));
+
+            if (products.contains(product)) {
+
+                productService.deleteProductById(product.getId());
+
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+        }
+        throw new ResourceNotFound("Not found products with the username " + username);
     }
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
